split code for 'new' state into new file
